WIP Hacking on getting neuvector click2deploy working
Enable billing support by default in the embedded copy of the
neuvector/core under k8s/neuvector/chart/neuvector.

Add README-mpdev.md outlining the steps needed to simulate a Google
marketplace deployment using mpdev.  Also added a template that can
be used to generate the contents for the userinitcfg.yaml secret.

k8s/neuvector/schema.yaml has a first pass at specifying the values
for NeuVector, subject to change.

Added a deployerServiceAccount definition that enables the deployer
to be able to create CRDs. More work needed to add any require RBAC
settings to the deployerServiceAccount to allow it to complete the
deployment, as well as defining any serviceAccounts that are needed
by the running deployment.

Note that the activemq Click to Deploy implementation has been used
as the reference example for helm based application deployment, as
well as the sample-app, which is an envsubst based deployment; still
need to cleanup some remaining ActiveMQ related content and replace
with equivalent for NeuVector.

Currently need to override the definition of C2D_CONTAINER_RELEASE
in the Makefile to specify a hardcoded value as the NeuVector images
don't have the required env setting, C2D_RELEASE, set when they are
being built.

Trim down the apptest test suite somewhat and start to add NeuVector
specific tests.